zebra: fix arp-nd-redirect access-vlan lookup
Original commit:
zebra: multiple vlan aware bridge data structure and related changes

When MVAB support was brought in, the EVPN access-vlan lookup started
referencing the ifp of the bridge interface instead of the bridge_slave
ifp. However, when that commit was brought into the upgrade branch, this
call to the EVPN access-vlan lookup was mangled to use the old logic
(bridge_slave ifp).

This commit re-fixes the lookup so the bridge is used instead of the
bridge_slave.

Issue: #3408358

Signed-off-by: Trey Aspelund <taspelund@nvidia.com>